Initialize also with pre-existing data dir (#190)
Currently, Microshift checks for the non-existence of the data dir (/var/lib/microshift) to infer that
it runs for the first time and thus should initialize certs, kubeconfigs, etc. This creates a problem
on systems where /var/lib/microshift already exists (e.g. for pre-loading of images or manifests).

Until we have cleaned-up initialization as part of the multi-node work, this patch introduces a work-
around that makes initialization conditional on the cert dir being present instead, which is more
specific.

Signed-off-by: Frank A. Zdarsky <fzdarsky@redhat.com>