Fix the flannel deployment
Flannel wasn't completely installed, since the host needs the flannel
binary in /opt/cni/bin for crio to find, and forceAddress is necessary
to make sure that the cni0 bridge will have the address requested
by flannel.

An alternative to this could be creating a flannel-cni rpm sub-package
to be installed on the host.

Signed-off-by: Miguel Angel Ajo <majopela@redhat.com>