Fix uid/gid setup and file capabilities
First, rpm requires the file attribute uid/gid to be symbolic
and not numeric. Additionally tar didn't store the caps
but cpio doesn't keep those inside rpm either.

So we create a list of file caps during tar creation, and then
we populate the %files metadata with %caps attributes so those
are re-installed during rpm unpack.

This is tested and working.