use release creation time as bindata timestamp
Update the rebase script to extract the creation date of the release
to use as the timestamp for the assets compiled into bindata.

Update bindata.sh to use the timestamp, if it is present.

Set the bindata_timestamp.txt content to the value for the current
release, 4.10.18.

Update bindata.go with the correct timestamp.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>