Refactor certificate creation to use structures instead of code
We allow specifying the certificate information using Go structures
instead of additional code that goes a certificate signing at a time
and checking for errors.

All of the certificate chains can be created by using the builder
patter of the "certificateChains", where all the signing information
can be injected, and once everything is done, calling "Complete()"
attempts to build all the certificate chains specified.

The "CertificateChains" structure, which is a result of the previous
chain completion allows querying for certificate PEMs in a directory-like
manner.

Each of the signers provides information about its directly-signed
certificates and its sub-CAs. Signers don't allow retrieving their
keys directly as those should never be needed to be extracted anywhere
in the code.