Create local APIServices for formerly-openshift-apiserver CRDs.
Certain resources (including RoleBindingRestrictions and
SecurityContextConstraints) belong to an API group served on OpenShift
by openshift-apiserver but have been migrated to CRDs. The
kube-apiserver on OpenShift carries a series of patches to ensure that
they coexist politely with openshift-apiserver. On MicroShift, the
group discovery handler for these types is never registered in the
aggregator (expecting openshift-apiserver to do it).