init: use a proper cn for the aggregator client certA
The current cert creation tooling might be confused if we did not
provide a service name. I'm supplying "remove-this" as this code
should be refactored soon.A

The current cert creation tooling might be confused if we did not
provide a service name. I'm supplying "remove-this" as this code
should be refactored soon.