USHIFT-573 avoid NodeName changes in running systems
NodeName changes are dangerous to running systems (attached storage,
running pods, etc, need to be relocated).

This change persists the NodeName in /var/lib/microshift/nodename,
and makes sure it's not modified across runs.

If the NodeName was auto-generated from hostname it will write
a Warning, and continue with the old NodeName.

If the NodeName is explicitly configured, it will fail with an error.