NP-603: use system MTU as default pod MTU
OVNKubernetes builds overlay networking implementation based on geneve
tunnel which adds extra header for packets coming from overlay (Pod) to
underlay network. However, the geneve tunnel is only useful for
cross-nodes communication and not required for microshift single node
deployment. This PR leverages the single-node change[1] in OVNKubernetes
and updates the default pod MTU to the same size as default uplink
MTU.

If system MTU is not detectable, it falls back to use 1500.

[1]: https://github.com/ovn-org/ovn-kubernetes/pull/3290