USHIFT-716: KAS advertise address match the service network ip
Adds an internal IP to kubernetes service. Pods will use an internal
IP to connect to KAS, while external clients will use the node IP.
Solves ambiguity problem when selecting which certificate to return
to a client, based on the destination IP (KAS).
Configures the service IP to the loopback interface so that the
endpoint slice for kubernetes service matches the same IP.