Apply service ca-bundle directly to ingress configmap
The data of openshift-ingress service-ca-bundle configmap
is erased everytime when microshift starts since the
configmap is applied without data. This results in delay
in starting router-default deployment pod after reboot
as it has to wait for service-ca controller to inject
the ca-bundle again. This commit applies service-ca-bundle
configmap directly with data generated in service-ca dir.

Signed-off-by: Zenghui Shi <zshi@redhat.com>