USHIFT-936: refactor reading and validating of config
This change streamlines the reading and validation functions so that
it is easier to get a Config instance with the defaults or with valid
values read from a file.

A new convenience function for loading the active configuration,
ignoring a missing configuration file, is added.

The k8s YAML parser is used to read the config file, instead of the
default parser, to ensure support for `json:"-"` directives to ignore
fields in the input.

The unit tests are updated to parse input config instead of using a
data structure. This more accurately simulates what happens when the
real program reads the config if only part of the file is populated,
and clarifies which part of the config each test scenario is
checking. It also allows for more, smaller, scenarios to isolate
configuration validation and defaulting failures.

The validation for node name changing is only relevant when starting
the server, so that is moved to the run command implementation to make
the unit tests for config easier to understand.

The show-config command implementation no longer needs to map between
multiple configuration data structures, so streamline that code.