OCPBUGS-8329: update sysconfwatch-controller logging
There is no need to log the gateway IP address every time we look it
up, so only log it when it changes. This is somewhat redundant to the
logic in the controller itself, but the IP returned from GetHostIP()
is not always the gateway IP and GetHostIP() may have other callers.

This change also rewords another log message related to how the host
IP is being selected to make it clear what effect the failure to find
a default route has on the outcome.