OCPBUGS-8277: Use internal IP for apiserver endpoint
Due to the support for IPs in certificates introduced in #1298, the
apiserver IP is configured as a secondary address in the lo interface.
A VIP is configured by ovnk redirecting 10.43.0.1:443 to 10.43.0.1:6443.
6443 is the port where apiserver listens in the host.
10.43.0.1:443 is used by all pods using client go, as it is computed
from the env vars we can find in any pod.
If a host network pod or any other tool in the host tries to reach the
apiserver by using 10.43.0.1:443 the address is not translated to the
endpoint, it tries to contact 10.43.0.1:443 which is not the apiserver
but the router. This change computes a new IP endpoint in the next
available /32 subnet from the service IP to ensure ovnk does not
interfere.