USHIFT-949: cache getAllHostnames
getAllHostnames is meant to run once on startup to help populate
configuration. In the test suite it may run many times, and it
occasionally blocks looking up the FQDN for hosts. This change caches
the first set of results to avoid blocking on future calls and to make
the tests run more quickly. This change is safe because MicroShift
should be restarted (automatically or by the user) if the network
settings of the host change in a significant way.