OCPBUGS-10766: Change apiserver preferred addresses
In order to avoid DNS configuration unless there is no other way, try
using external and internal IPs to reach kubelet from apiserver. This
allows having whatever hostname+domain name combinations without trying
to resolve the name externally.
To allow this we need to include the node IP in the kubelet certificate
as this is what the apiserver will target.