USHIFT-950: add verification step to ensure etcd/vendor is kept up to date
As we make changes to code in microshift that is vendored into etcd, we
need to ensure that we also update the vendored copy at the same time.

This change adds a Makefile target `vendor-etcd` that updates the
`etcd/vendor` directory and another target `verify-vendor-etcd` that fails
if there are any files added or modified after running `make vendor-etcd`.