support ssh agent with robotframework
On a local developer system it is more likely that the environment will
have ssh-agent configured than that there will be a passwordless ssh key
file available. This commit adds a resource file with keywords for logging
in and out so that if SSH_PRIV_KEY is not set the login is via the agent.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>