USHIFT-1238: fix indentation of PR messages
Using an inline f-string to build the message means that the dedent()
call is applied to the full text of the message, rather than just the
template. The results therefore depend on the messages in the
changelog, and come out misaligned. This change turns the dedent()
results into a template string and fills in the template separately so
that the results are more appealing.