USHIFT-1286: Use exit code 1 in greenboot failures
All MicroShift failures return 1 except for the service not being able
to start. This happens because it shortcuts error checking by sending
a SIGTERM to itself so that it exits immediately. The default handler
for this signal returns exit code 143, whereas all the other known
failure conditions return 1.
When testing this behavior the greenboot script is not consistent when
something fails. This change creates a specific handler for TERM so
that the exit code is also 1.