consider filepath as not existing even if exists but is empty
Before the prerun, MicroShift creates a .nodename file in the data
directory. To not interpret existence of the data dir with just
the .nodename as "MicroShift already ran on this device",
we need a more thorough check.

Extending this philosophy to files, will allow us in testing to
create empty health.json that will be considered as "no health info"
but writable by green or red script after adding immutable attr
too backups directory.