OCPBUGS-15740: Fix show-config command output
Command `microshift show-config` shows the combination of what's in
`/etc/microshift/config.yaml` and the defaults that are not in the
file.
In the event of a hostname change this file does not show the correct
hostnameOverride and subjectAltNames, as they are changed right before
starting microshift to preserve the original node name.
This fix runs the same logic when showing the configuration or when
starting microshift.