USHIFT-1624: quiet logging about node IP
The GetHostIP function is used repeatedly, which means we need to be
careful about how much it logs. This change adds a global state
variable to track whether the hard-coded node IP has been successfully
found, so that we only log that we are trying to find it if it has not
been sought before or if there is an error finding it.