healthcheck deployments and daemonsets
Instead of checking individual pod state, including restart count,
check the ready status of the deployment or daemonset that controls
the pod. Let the controllers for deployment and daemonset be
responsible for getting their pods into that ready state.

Because the old implementation applied the timeout to several
different phases, introduce a retry loop for each namespace to allow
each deployment or daemonset several attempts to get to ready.