shellcheck: disable rule that looks for unreachable code
The unreachable code detector is confused by the function in
packaging/greenboot/microshift-running-check.sh that is registered as
a trap to call exit when something goes wrong. This change disables
the rule globally instead of updating that script to add a directive
to disable it for every function in the script.