Pruning old backups: don't remove manual backups
Verifies if name of a backup is matching schema of automated backup.
Manual backups that start with `X.Y.Z_` or any other manual backup
with underscore in its name should not be subject to automated
backup pruning.

Update RF test to verify that pruning is limited to auto backups