More changes to improve UX
Changes prioritize console output over efficiency,
for example backupPathToStorageAndName will run two times:
as a check and to use the result.

Preconditions run as part of PersistentPreRunE instead of RunE
to avoid printing errors with logger which negatively impacts
the readability.

Restoring a backup is now allowed when microshift.service is failed
as it might be a way to get MicroShift to a healthy state.