USHIFT-1638: fix current release package source template
At the start of a release cycle there are no EC or RC builds for the
current release, so we should skip adding the package source. Package
sources are skipped if their template produces an empty file when
rendered. The crel template was producing a file with 1 blank
line. Fix the whitespace marker at the end of the template to avoid an
extra blank line.