USHIFT-1638: allow repeated configuration of logging
Tell the logging code that it's OK to receive reconfiguration
instructions unless those instructions are different. This overrides
the default behavior of erroring out if any instructions are received
a second time. That default behavior causes issues with embedded
components that all use the same logging code and are not expecting to
be compiled together into the same process, as they are in
MicroShift. See comments in
k8s.io/component-base/logs/api/v1/options.go for details.