USHIFT-1638: rebase with kube 1.28 update (#2450)
* update last_rebase.sh

* update changelog

* update microshift/go.mod

* update microshift/vendor

* update etcd/go.mod

* update etcd/vendor

* update component images

* update manifests

* update buildfiles

* add patch to reintroduce resourcequota.NewMonitor

See https://github.com/kubernetes/kubernetes/pull/120777

* apply resourcequota.NewMonitor patch

* update location of ServiceIPRange

The definition of ServiceIPRange changed in upstream kubernetes, so
update the way we import it.

* remove trailing whitespace in build_images.sh

* USHIFT-1638: fix current release package source template

At the start of a release cycle there are no EC or RC builds for the
current release, so we should skip adding the package source. Package
sources are skipped if their template produces an empty file when
rendered. The crel template was producing a file with 1 blank
line. Fix the whitespace marker at the end of the template to avoid an
extra blank line.

* USHIFT-1638: fix detection of empty blueprint file

Use the right local variable for the file to check for content to
detect that a template for a blueprint rendered empty.

* when crd manager receives cancel notice return the error

* USHIFT-1638: allow repeated configuration of logging

Tell the logging code that it's OK to receive reconfiguration
instructions unless those instructions are different. This overrides
the default behavior of erroring out if any instructions are received
a second time. That default behavior causes issues with embedded
components that all use the same logging code and are not expecting to
be compiled together into the same process, as they are in
MicroShift. See comments in
k8s.io/component-base/logs/api/v1/options.go for details.

---------

Co-authored-by: ci-robot <ci-robot@openshift.io>