restructure junit in scenario.sh
We need to collect junit output for multiple iterations of calling
scenario.sh for the same job. We could use different output files, but
that would mean complicating when we close the XML data
structure. Using separate files is simpler. The output file has to be
called 'junit.xml', so we want to put the files in separate
directories. This commit adds a "phase" directory for each scenario to
hold the separate junit files.