USHIFT-2105: cert rotation test wait for greenboot
This test changes the time immediately after starting. If the test
is executed when the system is running greenboot (from a reboot, for
example), the time warp makes it fail with fake timeouts, as time
jumped to the future. Depending on the test execution order this
might happen randomly.
Force waiting for greenboot before starting to avoid glitches.