remove FDO scenario
The original intent of the scenario was to test going from a base
image without MicroShift to an image that failed to come up, roll back
to the original image, then upgrade again to an image that worked. The
test was ensuring that MicroShift deleted the database during that
second upgrade so that any corrupt data would not cause an issue with
the new rollout. We subsequently removed that behavior, and updated
the FDO scenario so that it deploys 2 images that both fail. We don't
consider that a useful scenario.