USHIFT-2382: Increase timeout for osconfig periodic scenario
The osconfig suite takes a long time because of the restarts tests.
Reducing the amount of restarts has been discouraged, so instead we
increase the timeout for the whole test, as it was too close to 1h
and it flaked sometimes.
A different suite partition for execution should be studied as a
long term alternative.