do not require repo to be enabled to determine microshift versions
The `dnf repoquery` command does not require a repository to be
enabled in order to check its contents. Since we use the command to
look for old versions of MicroShift, in repositories that might not be
enabled, we should not require the repo to be enabled before
performing the query.