USHIFT-1349: automatically track pre-release versions in release note script
We build a default list of versions to scan using the current version
and the previous version. This assumes the script is run out of the
main branch where we can find the most current version under
development. During the period where 4.n is being developed, 4.n-1 may
still be producing only release candidates, so that scanning those 2
versions give us the 2 most recent candidates for having no
releases. During the period where the main branch has not landed a
rebase to update the version and both main and the pre-release branch
have the same version, we will end up scanning for EC or RC releases
of 4.n-2, but that's