OCPBUGS-32946: Default to watch all interfaces and IPs for router listenAddress
During first boot MicroShift will start scanning all IP addresses in the host to
setup the router load balancer service. When this happens, not all interfaces
may have an IP yet, leaving them out until the next microshift restart.

By leaving the defaults empty we allow the IP address subscription handler to add
anything that is not forbidden automatically. If this list is fixed with the IPs
and interfaces that were present only when MicroShift started, we might miss out
on some of hte listening IPs unless the user restarts.