USHIFT-4757: Skip installing gateway-api as optional rpm
Gateway API is a dev preview future, meaning there are no guarantees on
testing/docs.
When running conformance there is one issue with terminationMessagePolicy
values in istio pods that can not be overridden by the operator with current
options/configuration.
The RPM is installed using the common script configure-vm.sh which takes all
optional packages and installs them. Since this feature is not ready yet
for conformance, and the proper fix depends on another team we are disabling
the install temporarily.
Other e2e tests will keep testing the feature, as it is included in some of
the scenarios.