Upgrade openvswitch package version to 3.3
Open vSwitch 3.1 is EoL upstream and so the support of this version
in FDP is limited.  To ensure timely bug fix delivery for 4.14+ we
need to upgrade to the current LTS stream, which is Open vSwitch 3.3.
It will be fully supported for much longer.

This change follows a corresponding move of RHCOS and OVN-Kubernetes
to the same OVS 3.3.  MicroShift should stay in sync.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>